add container.jsx component and footer with social media links using bootstrap
