+++ src/components/footer/footer.module.css
@@ -2,8 +2,23 @@
   color: #2c3e50;
   width: 100%;
   text-align: center;
-  padding: 10px;
+  padding: 20px;
   position: relative;
   bottom: 0;
   left: 0;
+}
+
+.socialIcons {
+  margin-top: 10px;
+}
+
+.iconLink {
+  color: #2c3e50;
+  font-size: 24px;
+  margin: 0 10px;
+  transition: color 0.3s ease-in-out;
+}
+
+.iconLink:hover {
+  color: #f39c12;
 }
