+++ src/index.css
@@ -83,6 +83,16 @@ li {
   font-weight: bold;
 }
 
+.new-todo-page {
+  text-align: center;
+  padding: 20px;
+}
+
+.logo {
+  width: 50px;
+  margin-bottom: 20px;
+}
+
 @media (prefers-color-scheme: light) {
   :root {
     color: #213547;
